PL-1452: Add rule to limit output rows per memory sink where appropriate
Summary:
Depends on D3495. This rule adds a limit above memory sinks when the node above it isn't a limit.
If the node above it is a limit, it edits the limit where appropriate.
Next will be to hook this into the analyzer.

Test Plan: added tests

Reviewers: philkuz, jamesbartlett, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1452

Differential Revision: https://phab.corp.pixielabs.ai/D3496

GitOrigin-RevId: 8416c8cd9037895f877d22a4a3abeaa5ac4ec8dd