[PL-1438] Add UDF to map Pod ID to Node Name (px.pod_id_to_node_name(df.pod_id))
Summary: Network stats doesnt have UPID, so need to be able to map to the node name from pod_id. I built a weird join to make this work, but I'd like to avoid doing that in the future.

Test Plan: Added a test, tried in practice as well

Reviewers: nserrino, jamesbartlett, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PL-1438

Differential Revision: https://phab.corp.pixielabs.ai/D3512

GitOrigin-RevId: 7bf27ab38922f348789c0676b70c06f68db1759f