Fix unexpected govet issues by moving exported vizier udfs into funcs/export dir.
Summary: Depends on D3526. had a govet issue because gazelle liked to combine the cpp with the go code. Moved Vizier UDFs export into a child directory to make this more isolated and fixed the govet issue (hopefully, it's honestly unclear why this took so long to get triggered)

Test Plan: tests pass after the reorganization, hopefully govet likes this

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3529

GitOrigin-RevId: dba0b95f60c32a2b4c94f9e1835f1921bda74be2