[PL-1445] Allow aggregate expressions to be empty AKA use agg as a hash set on groups.
Summary:
The planner optimizer will remove columns if they are unused down the line. There are some cases where we might want to group by without caring about the output expression. This allows that to happen, meaning we don't have to make weird hoops to make sure we don't run into execution issues.

I'm worried this isn't a clean approach. Alternatively, we could just have a separate Hash function call that we hit whenever ConsumeNextImpl is called for an AggNode that has no values.

Test Plan: created a test that used to segfault, fixed the segfault

Reviewers: zasgar, nserrino, #engineering, michelle

Reviewed By: #engineering, michelle

Differential Revision: https://phab.corp.pixielabs.ai/D3531

GitOrigin-RevId: 6defc931dbd5651d713e8890d712ae8a86395c40