Stirling/Cassandra: Hook up CQL Tracer into Stirling - Part 1
Summary:
Hooking up CQL tracer. This is mostly following the pattern of the other protocol tracers.

Note that this doesn't fully enable the CQL probing yet. Change to the BPF code is required to set kProtocolCQL for that to happen. That is coming in Part 2.

Test Plan: Added a basic test to socket_trace_connector_test

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-711

Differential Revision: https://phab.corp.pixielabs.ai/D3536

GitOrigin-RevId: 0295fae56ec601bdfeef4bb8321344de72877dc6