PL-1497 part 2: Update filter column pruning logic to use column selection in output to fix bug
Summary:
We had a problem with filters where if you had two filters in a row, where both filters used columns for the filter condition
that were not used downstream anywhere, you would end up with misaligned output relations. We don't need to output columns that are
only used for evaluating a filter condition, so this diff updates the logic to take advantage of column selection in the filter node.

Test Plan: added

Reviewers: philkuz, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1497

Differential Revision: https://phab.corp.pixielabs.ai/D3551

GitOrigin-RevId: 070a55f15d43b39c3b59df403b47f6095ab590b3