Exec::Stdout() bug-fix
Summary:
Was using this in a cassandra test and ran into a bug in which Stdout was stuck in an infinite loop.

The bug occurred when 0 bytes were read (end-of-file), since no -1 would be returned to break the loop.

This version will break anytime the expected number of bytes is not read.

While at it, I also added an ability to return errors.

Test Plan: Tested it manually.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3558

GitOrigin-RevId: d47b9a2209b98f2864c346284d87c0d0f4ec681a