PL-1433: GRPC sources should get removed from the router when they go stale, even if query fails.
Summary:
Previously added a fix to remove GRPC source pointers from the router before those pointers go stale.
The previous fix was the right direction but it didn't cover removing them during failure cases for the exec graph.
This change removes those nodes even if the Init or Execute calls on execution graph fails.

Test Plan:
ran on the failing query for PL-1497 in a branch that preceded the recent fixes for PL-1497. Ran many times
because the failure is nondeterministic.

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1433

Differential Revision: https://phab.corp.pixielabs.ai/D3578

GitOrigin-RevId: 01038edd55c9dc32e549420fa938566ec50c6c1a