Make sure the deploy instructions info are up to date.
Summary:
When we added apollo cache persist, ALL the cloud GraphQL data was cached in localstorage, which means the data was fetch once and never again.
This makes the request for CLI info 'cache-and-network', which always makes a network request and keeps the cache consistent with the server.

Test Plan: tested manually

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PL-1503

Differential Revision: https://phab.corp.pixielabs.ai/D3589

GitOrigin-RevId: 29bb04c43ff8ecbf14b86d35d683cfa9f89148b2