Add arc unit check generated grpc_web_library files have been updated
Summary:
now that we have pl_grpc_web_library, we need to check the generated files have been updated in the UI directory.
the process is fairly involved. for each proto file, we need to check if it is a dependency for a pl_grpc_web_library target. to do this, we need to
search the BUILD.bazel files for the pl_grpc_web_library name, run a bazel query to get all of the deps of the target, and check that the current proto file matches any of the deps.

Test Plan: ran arc unit and tested a few cases

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3590

GitOrigin-RevId: 7c9379d8983de2f6e1cbd59f041c44d2e5071039