BPF hackery to bump up MAX_MSG_SIZE.
Summary:
This diff increases the MAX_MSG_SIZE to 30KiB.

While it might not look like much code, this was a long and hard-fought battle with the BPF verifier.

In the end, figured out that clang was actually getting in my way of sending hints to the verifier, so used some inline assembly to trick clang into not optimizing away my code.

Note that setting MAX_MSG_SIZE to 32KiB is not a straightforward option because the BPF map doesn't support an object that large.

Test Plan: Existing tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3600

GitOrigin-RevId: 8f3013f8bd5f6daa265c1bba23b52c9d04a797d4