CQL Tracer BPF Test
Summary:
This test starts a cassandra server in a container, and then runs some commands on it.

Meanwhile, stirling is running, and should capture the messages.

Currently only checking for client-side capture. Also need to support server-side capture, but BPF is causing issues there.

Test Plan: This is the major end-to-end CQL tracing test.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-711

Differential Revision: https://phab.corp.pixielabs.ai/D3587

GitOrigin-RevId: 55e426b8e0560b4130bd8d677f61e58f326947d3