Refactor metadata handler to allow subscribing to metadata updates
Summary:
before, we were only sending metadata updates to agents. we now want to send metadata updates to the cloud as well, so some of this logic needed to be cleaned up.
the metadata handler now has subscribers to its updates. this way the metadata handler doesnt need to know anything about agent queues/the message bus.

Test Plan: unit test, ran in skaffold and everything works as normal

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3620

GitOrigin-RevId: 5c07aebd2e069bb138b36449535568c95a551d73