CQL Seg-fault fix
Summary:
Seg-fault was due to a bad CQL protocol inference, which meant the frame that was parsed included garbage, which was interpreted as a frame with negative length.

For robustness, added checks to length and protocol version to make sure they're valid. Otherwise return kInvalid instead of seg-fault.

Test Plan: Test cases added.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1504

Differential Revision: https://phab.corp.pixielabs.ai/D3643

GitOrigin-RevId: 1f2083c5aea784dccd653598db896533a872a16f