EnumCast
Summary:
A wrapper around magic_enum::enum_cast, that converts the value to an error if the value is not cast-able. This is meant for use with PL_ASSIGN_OR_RETURN. Otherwise, it is simpler to just use magic_enum::enum_cast directly.

No use cases yet, but some are coming soon.

Test Plan: Test case added.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3672

GitOrigin-RevId: 47c53987ad2dbd5c0fb6d5eda14c298a23b6b344