CQL: Support Batch requests
Summary:
This parses the last remaining CQL command: BATCH.

Depends on D3659.

Test Plan:
Unfortuantely, I have not been able to generate a BATCH frame through cqlsh (even when using the BATCH command), so I have no data to test this on.
BATCH in cqlsh results in a QUERY frame with a BATCH command inside. Makes me wonder if this part of the binary protocol is even in use in the wild.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-711

Differential Revision: https://phab.corp.pixielabs.ai/D3660

GitOrigin-RevId: daacfa35784766cff7c785561c9ff3a70ca55af4