Add PrevResourceVersion to update
Summary:
when sending out an update, we need to know what was the previous update that was sent out so we can determine if any updates have been lost.
we will soon want to do a similar metadata update process for the agents, so I made the interface for getting resourceVersions/prevResourceVersions for a general "subscriber", where a subscriber is cloud or an individual agent. individual agents may have a different chain of updates, so we would need to keep a separate mapping of rvs to prevRvs for each agent.

Test Plan: unit tests, ran skaffold to make sure everything still runs

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3690

GitOrigin-RevId: 8abe181d67d9297cdc14872ec9169f0c033fec85