Reduce the size of etcd and nats to 1 pod
Summary: This just updates the schemas. If we don't want this in all cases (we do lose some etcd features and nats resiliency) we should implement the new RPC endpoint Zain discussed.

Test Plan: Tested that etcd and nats can be reduced to 1 pod and ran all the preset queries as well as the scenario.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3714

GitOrigin-RevId: 4cefa0e3dba7ce3ffa8f8597416987746270f5c4