Implement metadata queue reader
Summary:
The metadata queue reader reads from the DurableMetadataUpdate channel sends the metadata updates in order to the indexer.
if it finds that it receives an out of order metadata update, it requests for the correct metadata updates through nats.

currently there is no code which calls StartVizierUpdates, so nothing actually runs yet. that will come in a later diff.

Test Plan: unit tests, ran skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3712

GitOrigin-RevId: 00686154c5873c0a7ba3a0a107acb3420c664407