PL-1546 part 1: Support creating a new org from the Login handler
Summary:
We want to move to a model where users log in or sign up and don't have to worry about sites.
There are no subdomanins and the user doesn't have to explicitly name their org.
This is a big change so it is staged out in a few parts. This part adds support to the Login handler that performs
the org creation if the org doesn't already exist. However it requires a flag to be set that nowhere else in the code
sets to true, so it is only unit tested for now. the old way is still currently how it will work.
Next stage will be deleting the old way and pointing all sign up flows toward the login handlers.

Test Plan: added/edited tests

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1546

Differential Revision: https://phab.corp.pixielabs.ai/D3737

GitOrigin-RevId: 2223a1edc5e86a714fb90659e92d8b9f5b9e958d