PL-1546 part 2: replace create site with login handler, remove site manager
Summary:
Depends on D3757. This will break Pixie if we don't land this at the same time as UI changes,
so it shouldn't be landed alone. sites do not get created. the signup endpoint not the create site endpoint is used to make orgs.
 in subsequent diffs site manager will be refactored into project manager.
for now, it just got removed from the api environment.

Test Plan: existing, we also need to manually test this with the corresponding ui changes

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1546

Differential Revision: https://phab.corp.pixielabs.ai/D3740

GitOrigin-RevId: d94b7f942b8b5cd8f60ae16a8ad7780b1d46968a