Remove sites from UI
Summary:
We're removing the concept of sites from our product. This means that we're removing subdomains and the pages for creating sites from the UI.
all of the auth/login stuff is fairly nuanced, so please let me know if i'm missing something!
{F53666}

Test Plan: tested by running cloud locally. not fully able to test until backend changes are complete

Reviewers: malthus, #engineering

Reviewed By: malthus, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3756

GitOrigin-RevId: a8d6ac46dc42f660343f419019917b85cd9269fc