Add fluentbit annotations
Summary:
Depends on D3790.
Fluent-bit parses log messages into structured JSON using parser plugins.
We currently have two log formats in our code base so each deployment must specify to Fluent-bit which log format it uses.
Note that query-broker uses two log formats, which isn't that supported by fluent-bit. I added a bit of a hack to get it work in the config of fluent bit in a later diff. For the purposes of this diff just know that any deployment that has both glog and logfmt formats should specify its annotation as logfmt, and then the fluent bit config will take care of the glog format. Because of this its actually not strictly necessary to specify the annotation for any deployments which use glog, but I did it anyway for the sake of consistency.

Test Plan: Manually tested end to end.

Reviewers: #engineering, zasgar

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.c.pixielabs.ai/D3792

GitOrigin-RevId: cdf53800493a41d874ce77ba220945c34226ec34