PL-1558: Fix error message where user tries to sign up when their account already exists
Summary:
We were looking up by the wrong user ID (auth0 ID, not uuid). But it really should be looking up user by email.
If the user's email is present in the system then they shoudln't be able to sign up with that email.
so in this diff I added GetUserByEmail which mirrors GetOrgByDomain.

Test Plan: added tests, manually verified with skaffold

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1558

Differential Revision: https://phab.corp.pixielabs.ai/D3779

GitOrigin-RevId: d26878abba305e34d431ff8065c2c9aac7623257