PL-1431 part 1: Add ASTVisitor as a property of QLObject
Summary:
This is a bit deep in the yak shaving, but this diff is part of a refactor that includes the following improvements to the planner:
1. Remove CollectionIR/ListIR/TupleIR and consolidate all lists into CollectionObjects
2. Remove need for both Process() and ProcessData(), consolidate to Process()
3. Remove hack where all FuncObjects get Call() invoked on them when they are an argument to another node in order to produce an IRNode for the aggregate functions. AggHandler will now perform Call() itself.

In order for the AggHandler to call Call() itself, it needs the ast_visitor in order to populate the func with the default arguments.
Now any QLObject that may contain a FuncObject that can be Called needs to have a reference to the ASTVisitor in order to compute
the default arguments. So, it seems cleanest to make ASTVisitor* a property of QLObject, similar to how IR* is a property of all IRNode*s.

Test Plan: existing

Reviewers: philkuz, jamesbartlett, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1431

Differential Revision: https://phab.corp.pixielabs.ai/D3815

GitOrigin-RevId: f748fa7d863714fcb82d253467b3141eba6e0de9