PL-1502: support joins like merge(...left_on='a', right_on='b')
Summary:
Depends on D3820. D3820 actually introduces this behavior but I added a test for it here.
Now like drop and union, you don't have to pass a list of key columns as an array if there is only one
key column pair in the join.

Test Plan: added

Reviewers: philkuz, jamesbartlett, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1502

Differential Revision: https://phab.corp.pixielabs.ai/D3825

GitOrigin-RevId: 47425df542caa5b9423647b7c986c850b4fa6c50