PL-1431: Change CollectionObjects to be an array of QLObjectPtr
Summary:
This is part of the refactor for PL-1431, which is eventually getting rid of CollectionIR objects.
This diff changes the internal structure of CollectionObjects to be an array of QLObjectPtrs,
rather than a single CollectionIR.
This also allows us to remove the hacky calls to Call() on all funcs in the ASTVisitor,
because after D3815 can now just invoke Call() directly in the AggHandler where it's needed.
I also did some cleanup unifying how we process arrays/strings to use some of the helpers to reduce boiler plate.

Depends on D3815.

Test Plan: existing

Reviewers: philkuz, jamesbartlett, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1431

Differential Revision: https://phab.corp.pixielabs.ai/D3820

GitOrigin-RevId: 8a6677e65f65fcbc71adbe88afe8585212a55000