Expanding MySQL container test coverage
Summary:
Add more commands in the script.

Unfortunately, the PREPARE/EXECUTE commands translate to COM_QUERY instead of COM_PREPARE and COM_EXECUTE in mysql cli.

Will have to find another way to make these even more stressful, but that'll be a separate diff.

Test Plan: This is expanding the existing test.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3806

GitOrigin-RevId: f958451ec289ff7a2581b722d1d5b62a802f21d2