Update Console view to use new script bundle
Summary:
we have a new script bundle which describes all of the example pixie scripts. we should use this instead of the preset queries.
we also need to update live view to have a dropdown of these scripts until autocomplete is done. this will be in a followup diff.

Test Plan: ran webpack

Reviewers: malthus, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3844

GitOrigin-RevId: 5ae56e9f7bcdd46ab2ef4abcab00b19d204b8878