Adding log-collector service.
Summary:
Depends on D3787.

Adds a log collector service on the cloud side.
This service listens on the NATS channel v2c.*.*.log.
It then pushes the log messages to the elasticsearch cluster.

Note that currently, on startup of Vizier the cloud-connector spams a whole bunch (on the order of 20000) of failed to connect messages, so the log-collector buffer fills up when a Vizier is first starting and will drop some log messages during startup.

Test Plan: Added dockertest to check that things actually get written to elastic. Also tested end to end with other components.

Reviewers: #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: philkuz, zasgar

Differential Revision: https://phab.c.pixielabs.ai/D3789

GitOrigin-RevId: 88eed2100b92bab79dd304a2b323dc72e71a3feb