Upload spinnaker config to repo
Summary: we need to backup our spinnaker configs in our repo. this means separating our secrets, which we encrypt in sops, then plugging them in later when we need to generate the config. this is done using a script called generate_config.sh

Test Plan: ran generate_config and check that the config it created is correct

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.c.pixielabs.ai/D3838

GitOrigin-RevId: b4d021ddf50b2adac94e4938d74f6664d50e2ebe