MySQL: Don't push mal-formed or partial records
Summary:
To avoid noisy MySQL records in the tables, we now avoid pushing records to the table unless they look clean.

Previously, we wanted to allow "partial" data to be robust to lost BPF events, but because MySQL has some false positive classifications,
it seems like a better strategy to push only the clean records.

Test Plan: Added a couple of test cases to show it works.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1563

Differential Revision: https://phab.c.pixielabs.ai/D3855

GitOrigin-RevId: 7725099f126d2bc97cea0561814522557e96a6a4