Move voyager to dynamic import
Summary:
Our bundles are huge, and voyager makes up a large part of it. it makes sense to move voyager to a dynamic import, since we only need it if the user has voyager.
for the other large modules, i'll need to try some other scheme.

this change requires changing our tsconfig version to esnext, which allows "import" which is used in the dynamic import. earlier this import was just getting changed to "require", which is not a dynamic import

Test Plan: ran webpack and tried it

Reviewers: malthus, zasgar, #engineering

Reviewed By: malthus, zasgar, #engineering

Differential Revision: https://phab.c.pixielabs.ai/D3870

GitOrigin-RevId: 58efad2e3f0b120fdde60da5a92f219c2000c324