Create VizObject to enable decorator
Summary:
Depends on D3878. VizObject is the basis for the decorator and has Vega function.
`px.viz.vega` takes a string config as an argument and then returns a function that the decorator parser
then calls with the function to wrap as the argument. The function is updated to have the vega spec.

Test Plan: tested in a later diff, might add one here if we want it.

Reviewers: nserrino, jamesbartlett, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.c.pixielabs.ai/D3879

GitOrigin-RevId: c1f0bf9a42563540266fede163ad93204009a2a6