Protocol Stitchers: return error count
Summary:
This is a foundational piece that changes the ProcessFrames() API of the protocol parsing pipeline such that it returns an error count in addition to the vector of records.

The plan is to use the error count to disable any connection trackers that are having too many errors.

Test Plan: TBD.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.c.pixielabs.ai/D3868

GitOrigin-RevId: 51828454ee2f6efce7cb7002818e87f576dfb0bd