ConnectionTracker: Disable after crossing stitcher failure threshold
Summary:
Similar to auto-disablement of a connection tracker when then frame parsing error rate is too high,
this diff adds auto-disablement of a connection tracker when the stitcher error rate is too high.

Test Plan: Added a couple of tests: one for the frame parsing case, and one for the stitching case.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.c.pixielabs.ai/D3876

GitOrigin-RevId: a45742d0d414db8c225629bf641e527e2dcd39d6