Dynamically import vega
Summary: in an effort to decrease bundle size, i'm following the example in https://webpack.js.org/guides/code-splitting/#prefetchingpreloading-modules where we do a dynamic import of the vega module (which is quite large). if the package hasnt yet been downloaded, we show a spinner. this should not show for too long for decent internet speeds.

Test Plan: checked tests still pass, ran webpack and tried it out

Reviewers: malthus, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.c.pixielabs.ai/D3898

GitOrigin-RevId: 89bf2e92007fde3fb326e09c504226e70f52108d