Stirling missing events: Remove problematic BPF map by using timestamps as generation numbers
Summary:
This is part of the fix for the missing events issue.

Root cause is BPF maps becoming full.

In this diff, we entirely remove a BPF map that would otherwise get full.

The purpose of this BPF map was to keep generation numbers of PID+FD combos,
but a timestamp works just as well. Only trade-off is that 64-bit number instead of 32-bit number.
But seems worth it to reduce complexity and avoid a BPF map issue.

Still more to come, as this wasn't the only map to become full...

Test Plan: Tested manually by running for ~60 mins on GKE.

Reviewers: yzhao, zasgar, #engineering, philkuz

Reviewed By: yzhao, #engineering, philkuz

Subscribers: philkuz

JIRA Issues: PL-1437

Differential Revision: https://phab.c.pixielabs.ai/D3903

GitOrigin-RevId: 4f63155cbd15760d735997e2d831939dfb322c8b