BPF: Fix conn_info_map resource leak
Summary:
This diff is another fix to address the issue where Stirling stops collecting events after 30-45 mins.

Root cause is BPF map resource leaks.

This particular diff addresses a leak in conn_info_map, where we don't delete the entry if we decide not to trace the connection.

Depends on D3903.

Test Plan: Tested on GKE running sock-shop. Requires a very long-running test to verify, so we need to give thought to a structured test.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1437

Differential Revision: https://phab.c.pixielabs.ai/D3911

GitOrigin-RevId: 6589df8df251680576ac131a6c63e85d4b6b201d