Add envoy to cloud
Summary:
We need to route grpc-web requests to the API service. this can be done using envoy, similar to our setup in vizier.
Unfortunately, there is one ugly part about these configs: the envoy yaml. there is a tiny part in the envoy yaml where we need to specify the api-service path which differs by just plc/plc-staging/plc-dev...
kustomize doesn't support variable substituion for configmap data: https://github.com/kubernetes-sigs/kustomize/issues/549 and they don't plan to
this means that the entire envoy config is replicated for each env, which is not ideal.

Test Plan: ran cloud and made sure a grpc-web request from the UI makes it to the API service

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.c.pixielabs.ai/D3909

GitOrigin-RevId: b5ee4b3255c5d1168c017f5ca51e7a1f099851f7