Change the console view to use grpc.
Summary:
Replace graphql calls with grpc to execute script. Refactor all the component in the console view to use the proto format of query results.

The only thing missing is the compiler errors, which would be added when the vizier side is fixed.

Test Plan: tested manually, fixed the existing tests

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3957

GitOrigin-RevId: 16374cc97f8838d9d3ae6146ae5403d334b7288c