Remove unnecessary extern C declarations
Summary:
extern C is about linkage/name mangling.

It isn't required for C and C++ to share structs.

Also, we never actually call a library from a .o file compiled with the C-calling convention.

The only thing we do is share structs from BPF to C++, but these shouldn't require extern C.

So we can safely remove it in a number of places.

Test Plan: Existing tests

Reviewers: yzhao, #engineering, zasgar

Reviewed By: yzhao, #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D3945

GitOrigin-RevId: 6a4568a80d89979242cce34329ae8e827e24d9c6