Test for BPF map leak
Summary:
This is a follow-up to D3925.

It adds a test to make sure we don't have any regressions.

Essentially, it creates server that leaks a BPF map entry when it is killed.

Then we monitor to make sure that the ConnectionTracker destructor eventually cleans up the BPF map entry.

Test Plan: This is a test.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1437

Differential Revision: https://phab.corp.pixielabs.ai/D3942

GitOrigin-RevId: 79c2698b406f5a6e71420917c729445a149be279