Remove BPF entries on ConnectionTracker destructor
Summary:
This is a safety net to help avoid leaking BPF maps.

Here we de-allocate BPF map entries when a connection tracker is destroyed.

Presumably, a close() syscall would cause the same thing to happen, but here also perform the delete, just in case.

This also helps in case a program was closed/killed without calling close on it's open connections.

Test Plan: TBD

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3925

GitOrigin-RevId: 4f30635721e1a820fb90c7382f1a84ce5a9dca50