Fix envoy config
Summary:
we need to fix the timeout and allow credentials for cookies.

we also need a separate yaml for dev vs prod/staging, because alpn options dont work in dev. hopefully we'll find a better fix for this once we figure out what is going on with the healthcheck

Test Plan: tested in skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3967

GitOrigin-RevId: 2d8d690c9458c39479f2d0c4ea765915d7f35902