Upgrade yarn/webpack and make build more verbose
Summary:
I tried to repro hang locally and could not do it. The jenkins hangs are happening on epollwait
which just means it's waiting for a file descriptor. The process is stuck, not doing doing anything
or consuming any CPU resources.

Hopefully the upgrade fixes the build. If not, we will have more debug info with the verbose build.

Test Plan: Jenkins

Reviewers: michelle, malthus, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3971

GitOrigin-RevId: 9a06aeadaca109abcaa2c2f1289f6e21278beaee