Add command for updating/getting cluster config in CLI
Summary:
we need a command for getting/updating the cluster config in the cli.
if the user just runs: `px config`, they will get a printout of their config which currently is just the string: PassthroughEnabled: true|false
If the user runs `px config --update --passthrough=true` it will update the passthroughEnabled config to true. if they just run `px config --update`, no values will change.
The reason I made the `passthrough` flag a string is because if it is a bool, a default value will be set and may update the user's config if they did not mean to

Test Plan: ran the command with get, verified that it got the correct config. ran the command for update, verified that the database value changed.

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3937

GitOrigin-RevId: dfa60366f2dd58503f9a447d5b27522467642959