Add CompilerError to vizierpb
Summary:
currently we are passing CompilerErrors as an Any in the vizierpb status. we shouldn't expose our internal protobufs to external users, such as the UI.
instead, create set of errors that we can return in the status.
a followup diff will update QueryBroker to correctly translate the internal version of CompilerErrors to the external version of CompilerError

Test Plan: n/a

Reviewers: philkuz, malthus, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3952

GitOrigin-RevId: 731b84aeebea9da4422a486bbe5a8702ca7653fd