Add passthrough in UI
Summary:
if passthrough is enabled, we should make the grpc-web request to cloud instead of to vizier.
since cloud dev (dev.withpixie.dev) is not behind a google cloud load balancer, which would redirect /pl.api to the correct envoy port, we need to manually direct to the correct port.

Test Plan: ran dev cloud in skaffold

Reviewers: malthus, nserrino, zasgar, #engineering

Reviewed By: malthus, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3913

GitOrigin-RevId: a1afa64a40cdee306d672c1194755ffaeef5a042