Remove the old vizier graphql health check.
Summary:
This is the last piece of graphql API on the vizier side.
It shows the loading screen when vizier is not connected, if the UI somehow disconnect from vizier it will only disable the execute button instead of switching to loading screen.

Maybe we can add a timeout later to if the UI is diconnected for a long time.

Test Plan: tested manuall

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3984

GitOrigin-RevId: 42827083c24755e48b4368f986c4b0f2fe9cc4b2