profile service datastore: GetUserByEmail should return nil if there are no rows returned
Summary: profile server.go expects that when no email exists for any registered user, GetUserByEmail will return nil for user. We currently return an error for no user in db with that email. This non-grpc status error gets pushed upstream and confuses the caller. What we really want is to push a not found status error which can then be correctly interpreted. So, correct datastore to actually return nil for the "no user" case.

Test Plan: edited existing

Reviewers: michelle, #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D4013

GitOrigin-RevId: d8df2cfccef9712f40c222a24658821475655921