Remove voyager from UI
Summary:
voyager is pretty big and making our webpack builds really slow/hang. we tried moving it to an external dependency, but that
process is fairly complex and probably not worth our time if we don't even use voyager anymore. solution for now is to remove it unless
we have a need for it in the future.

Test Plan: ran webpack, make sure everything works as expected

Reviewers: malthus, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4027

GitOrigin-RevId: 668d533fc42debe2519b9967de5ce7427d2247e0