Add slackin proxy server
Summary:
This adds the proxy server for slakin running in the namespace slackin on the prod cluster.

Ideally this will not be in the same proxy as our app, but seems like low ROI to make this be in a seperate server,
backed by a different load balancer.

Test Plan: nginx starts :)

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PL-1615

Differential Revision: https://phab.corp.pixielabs.ai/D4041

GitOrigin-RevId: 75c1fd7fbb071470558685c7699670f88e7620e9