PL-1642: Memory sink nodes should write out 0-row RowBatches when that RB is eos or eow.
Summary:
We used to skip writing out 0-row RowBatches because it seemed like there was no real use for it.
However, it turned out to be important so that queries that produce 0 rows will still return a 0-row, end of stream result.
PL-1642 says that 0-row output tables never set an end of stream marker so this diff causes them to be written out
so the 0-row row batch with the end of stream marker still gets produced.

Test Plan: existing/added

Reviewers: jamesbartlett, philkuz, michelle, #engineering, zasgar

Reviewed By: #engineering, zasgar

JIRA Issues: PL-1642

Differential Revision: https://phab.corp.pixielabs.ai/D4042

GitOrigin-RevId: 48492b5e42b29239cfde1eeb798dc30b1efec14a