Add script for reading from prod db
Summary:
we need a script for users to read from the prod db for debugging purposes.
we should give these users access only to the prod-ro namespace in the prod cluster. this should allow them access to exec into the db-reader pod in the prod-ro namespace
and read the secrets to log into the db.

Test Plan: ran it, but we will actually need to create some users with restricted prod access to make sure this actually works

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4049

GitOrigin-RevId: 26742f0a7111e076ea38f2ef2613c011dc15daf2