BPF tracing model: dedicate entry probes to collect args only (accept() and connect() probes)
Summary:
Basic idea is that the only purpose of an entry probe is to capture the arguments.
All actual processing should be done on the return probe, once we have both arguments and return value.

So the pattern becomes:
1) entry probe: stash arguments
2) return probe: unstash arguments
3) return probe: process data

This also consolidates all logic in one place, versus before where we had some logic in the entry probe and some logic in the return probe.
By keeping it all in one function, it makes it easier to follow.

Implemented this on connect() and accept() as a preview, but we can follow the same pattern elsewhere.

One potential downside is it might slightly increase the instruction count of the return probes, pushing us to the BPF instruction count limit.

Test Plan: Existing tests

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D4072

GitOrigin-RevId: 3740b7123b93c9cb0d42607eb6ad45d4c007baad